+++ src/commands/mod.rs
@@ -1,5 +1,6 @@
 mod forall;
 mod init;
+mod manifest;
 mod status;
 mod sync;
 mod topdir;
@@ -7,6 +8,9 @@ mod update;
 
 pub use forall::run as forall;
 pub use init::run as init;
+pub use manifest::add_run as manifest_add;
+pub use manifest::modify_run as manifest_modify;
+pub use manifest::remove_run as manifest_remove;
 pub use status::run as status;
 pub use sync::run as sync;
 pub use topdir::run as topdir;