+++ src/workspace/manifest.rs
@@ -6,6 +6,11 @@ use std::path::{Path, PathBuf};
 
 use serde::{Deserialize, Serialize};
 
+const DEFAULT_LOCATION: &str = ".";
+fn default_location() -> PathBuf {
+    PathBuf::from(".")
+}
+
 /// Represents the complete Chord manifest file.
 #[derive(Serialize, Deserialize)]
 pub struct Manifest {
@@ -28,10 +33,16 @@ pub struct Repo {
 
     /// The directory to clone the repo to, defaults to the directory where
     /// the Chord manifest is located.
-    pub location: Option<PathBuf>,
+    #[serde(default = "default_location")]
+    pub location: PathBuf,
 }
 
 impl Manifest {
+    /// Initializes a manifest struct with an empty repos vector.
+    pub fn new() -> Self {
+        Manifest { repos: vec![] }
+    }
+
     /// Opens and deserializes the manifest file into a Manifest struct.
     pub fn read(top_dir: impl AsRef<Path>) -> Result<Self> {
         let manifest_file =
@@ -47,6 +58,14 @@ impl Manifest {
         }
     }
 
+    /// Creates a manifest file of the current configuration, overwriting
+    /// any existing one in the provided directory.
+    pub fn write(&mut self, top_dir: impl AsRef<Path>) -> Result<()> {
+        let mut manifest_file = File::create(top_dir.as_ref().join("chord.yaml"))?;
+        serde_saphyr::to_io_writer(&mut manifest_file, &self)?;
+        Ok(())
+    }
+
     /// Updates the manifest with the revisions in the lockfile. This is a
     /// destructive action, meaning the lockfile gets emptied out into the
     /// manifest.
@@ -56,6 +75,84 @@ impl Manifest {
                 repo.revision = revision;
             }
         }
+    }
+
+    /// Creates a new repo struct and inserts it into the manifest
+    /// file. Repo name must not already exist in manifest.
+    pub fn add_repo(
+        &mut self,
+        name: String,
+        remote: String,
+        revision: String,
+        location: Option<PathBuf>,
+    ) -> Result<()> {
+        for i in 0..self.repos.len() {
+            if self.repos[i].name == name {
+                bail!("'{}' repo already exists in manifest", name);
+            }
+        }
+        self.repos.push(Repo {
+            name,
+            remote,
+            revision,
+            location: match location {
+                Some(location) => location,
+                None => PathBuf::from(DEFAULT_LOCATION),
+            },
+        });
+        Ok(())
+    }
+
+    /// Removes a repo with the specified name from the manifest,
+    /// returns error if repo was not found.
+    pub fn remove_repo(&mut self, name: String) -> Result<()> {
+        for i in 0..self.repos.len() {
+            if self.repos[i].name == name {
+                self.repos.remove(i);
+                return Ok(());
+            }
+        }
+        bail!("'{}' repo does not exist in the manifest", name)
+    }
+
+    /// Modifies the specified repo with the provided values. Returns
+    /// an error if was not found.
+    pub fn modify_repo(
+        &mut self,
+        name: String,
+        new_name: Option<String>,
+        new_remote: Option<String>,
+        new_revision: Option<String>,
+        new_location: Option<PathBuf>,
+    ) -> Result<()> {
+        for i in 0..self.repos.len() {
+            if self.repos[i].name == name {
+                if let Some(new_name) = new_name {
+                    self.repos[i].name = new_name
+                }
+
+                if let Some(new_remote) = new_remote {
+                    self.repos[i].remote = new_remote
+                }
+
+                if let Some(new_revision) = new_revision {
+                    self.repos[i].revision = new_revision
+                }
+
+                if let Some(new_location) = new_location {
+                    self.repos[i].location = new_location
+                }
+                return Ok(());
+            }
+        }
+        bail!("'{}' repo does not exist in the manifest", name)
+    }
+}
+
+/// Uses new to define default instance of struct.
+impl Default for Manifest {
+    fn default() -> Self {
+        Self::new()
     }
 }
 
@@ -70,7 +167,7 @@ mod tests {
             remote: String::from("https://example.com/repo"),
             revision: String::from(revision),
             name: String::from(name),
-            location: None,
+            location: default_location(),
         }
     }
 
@@ -122,5 +219,43 @@ mod tests {
         serde_saphyr::to_io_writer(&mut writer, &test_manifest).unwrap();
 
         assert!(Manifest::read(dir.path()).is_err())
+    }
+
+    #[test]
+    fn test_remove_repo_inexistent_repo() {
+        let mut test_manifest = Manifest {
+            repos: vec![test_repo("repo-a", "main"), test_repo("repo-b", "main")],
+        };
+        assert!(test_manifest.remove_repo(String::from("repo-c")).is_err());
+    }
+
+    #[test]
+    fn test_add_repo_already_exists() {
+        let mut test_manifest = Manifest {
+            repos: vec![test_repo("repo-a", "main"), test_repo("repo-b", "main")],
+        };
+        assert!(
+            test_manifest
+                .add_repo(
+                    String::from("repo-a"),
+                    String::from("url"),
+                    String::from("rev"),
+                    None
+                )
+                .is_err()
+        );
+    }
+
+    #[test]
+    fn test_modify_repo_inexistent_repo() {
+        let mut test_manifest = Manifest {
+            repos: vec![test_repo("repo-a", "main"), test_repo("repo-b", "main")],
+        };
+        // Arguments don't matter, repo won't be found
+        assert!(
+            test_manifest
+                .modify_repo(String::from("repo-c"), None, None, None, None)
+                .is_err()
+        );
     }
 }